refactor: extract common utils
