+++ apps/api/src/starknet/ipfs.ts
@@ -6,12 +6,10 @@ import L1AvatarExectionStrategyAbi from './abis/l1/L1AvatarExectionStrategy.json
 import { FullConfig } from './config';
 import {
   ExecutionStrategy,
-  ProposalMetadataItem,
   SpaceMetadataItem,
-  StrategiesParsedMetadataDataItem,
-  VoteMetadataItem
+  StrategiesParsedMetadataDataItem
 } from '../../.checkpoint/models';
-import { dropIpfs, getJSON, getSpaceName } from '../utils';
+import { dropIpfs, getJSON, getSpaceName } from '../common/utils';
 
 export async function handleSpaceMetadata(
   space: string,
@@ -158,68 +156,6 @@ export async function handleSpaceMetadata(
   }
 
   await spaceMetadataItem.save();
-}
-
-export async function handleProposalMetadata(
-  metadataUri: string,
-  config: FullConfig
-) {
-  const exists = await ProposalMetadataItem.loadEntity(
-    dropIpfs(metadataUri),
-    config.indexerName
-  );
-  if (exists) return;
-
-  const proposalMetadataItem = new ProposalMetadataItem(
-    dropIpfs(metadataUri),
-    config.indexerName
-  );
-  proposalMetadataItem.choices = ['For', 'Against', 'Abstain'];
-  proposalMetadataItem.labels = [];
-
-  const metadata: any = await getJSON(metadataUri);
-  if (metadata.title) proposalMetadataItem.title = metadata.title;
-  if (metadata.body) proposalMetadataItem.body = metadata.body;
-  if (metadata.discussion)
-    proposalMetadataItem.discussion = metadata.discussion;
-  if (metadata.execution)
-    proposalMetadataItem.execution = JSON.stringify(metadata.execution);
-  if (
-    Array.isArray(metadata.labels) &&
-    metadata.labels.every((label: string) => typeof label === 'string')
-  ) {
-    proposalMetadataItem.labels = metadata.labels;
-  }
-  if (
-    Array.isArray(metadata.choices) &&
-    metadata.choices.length === 3 &&
-    metadata.choices.every((choice: string) => typeof choice === 'string')
-  ) {
-    proposalMetadataItem.choices = metadata.choices;
-  }
-
-  await proposalMetadataItem.save();
-}
-
-export async function handleVoteMetadata(
-  metadataUri: string,
-  config: FullConfig
-) {
-  const exists = await VoteMetadataItem.loadEntity(
-    dropIpfs(metadataUri),
-    config.indexerName
-  );
-  if (exists) return;
-
-  const voteMetadataItem = new VoteMetadataItem(
-    dropIpfs(metadataUri),
-    config.indexerName
-  );
-
-  const metadata: any = await getJSON(metadataUri);
-  voteMetadataItem.reason = metadata.reason ?? '';
-
-  await voteMetadataItem.save();
 }
 
 export async function handleStrategiesParsedMetadata(
