+++ apps/api/src/starknet/utils.ts
@@ -16,13 +16,8 @@ import EncodersAbi from './abis/encoders.json';
 import ExecutionStrategyAbi from './abis/executionStrategy.json';
 import SimpleQuorumExecutionStrategyAbi from './abis/l1/SimpleQuorumExecutionStrategy.json';
 import { FullConfig } from './config';
-import { handleStrategiesParsedMetadata } from './ipfs';
-import {
-  Space,
-  StrategiesParsedMetadataItem,
-  VotingPowerValidationStrategiesParsedMetadataItem
-} from '../../.checkpoint/models';
-import { dropIpfs, getJSON } from '../utils';
+import { Space } from '../../.checkpoint/models';
+import { handleVotingPowerValidationMetadata } from '../common/ipfs';
 
 type StrategyConfig = {
   address: BigNumberish;
@@ -37,24 +32,6 @@ export function toAddress(bn: any) {
   } catch (e) {
     return bn;
   }
-}
-
-export function getUrl(uri: string, gateway = 'pineapple.fyi') {
-  const ipfsGateway = `https://${gateway}`;
-  if (!uri) return null;
-  if (
-    !uri.startsWith('ipfs://') &&
-    !uri.startsWith('ipns://') &&
-    !uri.startsWith('https://') &&
-    !uri.startsWith('http://')
-  )
-    return `${ipfsGateway}/ipfs/${uri}`;
-  const uriScheme = uri.split('://')[0];
-  if (uriScheme === 'ipfs')
-    return uri.replace('ipfs://', `${ipfsGateway}/ipfs/`);
-  if (uriScheme === 'ipns')
-    return uri.replace('ipns://', `${ipfsGateway}/ipns/`);
-  return uri;
 }
 
 export function longStringToText(array: string[]): string {
@@ -217,58 +194,6 @@ export async function updateProposaValidationStrategy(
       console.log('failed to handle voting power strategies metadata', e);
     }
   }
-}
-
-export async function handleStrategiesMetadata(
-  spaceId: string,
-  metadataUris: string[],
-  startingIndex: number,
-  config: FullConfig,
-  type:
-    | typeof StrategiesParsedMetadataItem
-    | typeof VotingPowerValidationStrategiesParsedMetadataItem = StrategiesParsedMetadataItem
-) {
-  for (let i = 0; i < metadataUris.length; i++) {
-    const metadataUri = metadataUris[i];
-    if (!metadataUri) continue;
-
-    const index = startingIndex + i;
-    const uniqueId = `${spaceId}/${index}/${dropIpfs(metadataUri)}`;
-
-    const exists = await type.loadEntity(uniqueId, config.indexerName);
-    if (exists) continue;
-
-    const strategiesParsedMetadataItem = new type(uniqueId, config.indexerName);
-    strategiesParsedMetadataItem.space = spaceId;
-    strategiesParsedMetadataItem.index = index;
-
-    if (metadataUri.startsWith('ipfs://')) {
-      strategiesParsedMetadataItem.data = dropIpfs(metadataUri);
-
-      await handleStrategiesParsedMetadata(metadataUri, config);
-    }
-
-    await strategiesParsedMetadataItem.save();
-  }
-}
-
-export async function handleVotingPowerValidationMetadata(
-  spaceId: string,
-  metadataUri: string,
-  config: FullConfig
-) {
-  if (!metadataUri) return;
-
-  const metadata: any = await getJSON(metadataUri);
-  if (!metadata.strategies_metadata) return;
-
-  await handleStrategiesMetadata(
-    spaceId,
-    metadata.strategies_metadata,
-    0,
-    config,
-    VotingPowerValidationStrategiesParsedMetadataItem
-  );
 }
 
 export async function registerProposal(
